+++ LinqToRest/ResourceRetriever.cs
@@ -7,7 +7,6 @@ using System.Threading;
 using System.Threading.Tasks;
 using JetBrains.Annotations;
 using Messerli.LinqToRest.Entities;
-using Messerli.ServerCommunication;
 using Messerli.Utility.Extension;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;